Moving code from BatchJoinProcessing header files to impl files
